+++ src/emulator/np/src/trophy/trp_parser.cpp
@@ -126,14 +126,14 @@ bool TRPFile::get_entry_data(const std::uint32_t idx, TRPWriteFunc write_func) {
     return true;
 }
 
-const std::uint32_t TRPFile::search_file(const char *name) {
+const std::int32_t TRPFile::search_file(const char *name) {
     for (std::size_t i = 0; i < entries.size(); i++) {
         if (strncmp(entries[i].filename.c_str(), name, strlen(name)) == 0) {
-            return static_cast<std::uint32_t>(i);
+            return static_cast<std::int32_t>(i);
         }
     }
 
-    return static_cast<std::uint32_t>(-1);
+    return -1;
 }
 
 } // namespace emu::np::trophy
